requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
pydantic==2.4.2 
httpx==0.23.3
fastapi-cors==0.5.2
openai>=1.14.3
python-multipart==0.0.20
ffmpeg-python==0.2.0
pydub==0.25.1
numpy==1.26.4